MVP - working conditional breakpoints!
While it does actually work, the API needs a serious redesign, since as
it stands, the lifetimes seem to "force" uneccessary copying when
evaluating agent expressions. See the `armv4t` example for details.
Basically, the root of the issue stems from the fact that when iterating
over the list of registered BytecodeIDs associated with a breakpoint,
you're calling a callback function which can modify `self`.
Unfortunately, the most "obvious" place to put the data structure that
stores BreakpointIDs is on `self` itself, which results in some spooky
lifetime issues.