MVP tweak - remove uneccessary clones in eval loop
while this does ostensibly fix the issues I outlined in my last commit,
it doesn't really "solve" the underlying issue (i.e: that the agent
shouldn't be owned by the target, given that it operates on the target).

This approach is _okay_, but given `gdbstub`'s otherwise great track
record of avoiding API footguns, this hack seems out of place.

I see two approaches:

1. leave the current API in with a big **WARNING: THIS IS UNSTABLE**
 warning. I mean, the entire API is unstable, but this API would be
 _extra_ unstable.

2. block this feature until a cleaner API design is found

Honestly, which option I'll go with is a product of how much time I have
to experiment + how long this feature would languish in its own branch.

My gut feeling is to just merge this implementation and break it later,
but we'll see...