remove eager self.current_mem_tid assignment
The reason this logic was in-place to begin with was to work around the
fact that the GDB client would send "qC" prior to "H". As it turns out,
it's totally fine to ignore the "qC" packet, which was removed from
`gdbstub` many commits ago. As such, this code just hung around for no
good reason, so it's nice to see it gone :)