rework resume actions + report step/cont-with-signal
this is a big change I've been meaning to do for a while.

The main issue with the old approach of passing the Actions as an
iterator directly to `resume` was that it made handling invalid actions
pretty painful, requiring the use of closure-captured error values, and
other nasty things.

Additionally, the old approach wouldn't "scale" nicely to support the
various optional resume actions (such as range-stepping or - down the
road - the "stop" action), requiring end users to match against
ResumeActions that GDB would never actually send. Not great.

This new approach is admittedly a bit clunkier for end users, requiring
them to track any thread-id specific resume actions themselves, but on
the upside, it really simplifies gdbstub's internal implementation, and
does intruduce some additional API ergonomics to the end user wrt. not
having to worry about unsupported resume actions.

* * *

Another improvement is that I've added a new set of ThreadID and IdKind
variants that don't suppor the `Any` variant. By using these new
variants while parsing VCont, all the logic required to reject `Any`
variants is pushed down the call-graph, back to when the packets are
being parsed. Very nice.

* * *

There's still some optimization work that should get done wrt. how the
deprecated 'c' and 's' packets are handled (i.e: instead of wrapping
them in an `Actions` and re-using the entire vCont codepath, streamline
things a bit), but broadly speaking it is a lot nicer now.

* * *

Along the way, I've also finally exposed the resume-with-signal actions.

Thanks to a GDB bug, targets _must_ support handling these cases. In
older versions of `gdbstub`, the library would internally error if the
GDB client ever sent a resume-with-signal, but now targets can at-least
get the chance to handle them if they want to.